requirements.txt
dspy-ai
groq
opencv-python
python-dotenv
reportlab
